+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Major
 - Restructures `src` files into `client` and `server` folders. [#220](https://github.com/n8rzz/atc/issues/220)
 - Updates Node to version 7.0.0 [#184](https://github.com/n8rzz/atc/issues/184)
+- Moves aircraft command logic from `AircraftInstanceModel` to new `AircraftCommander` class [#181](https://github.com/n8rzz/atc/issues/181)
 - Adds `spawnPatterns` to airport json and vastly simplifies aircraft creation. Work on this issue ended up resolving many other smaller issues listed below. [#243](https://github.com/n8rzz/atc/issues/243)
   - Restructure `Airport.departures` to utilize routes [#229](https://github.com/n8rzz/atc/issues/229)
   - Abstract inline fix object out of ArrivalBase [#56](https://github.com/n8rzz/atc/issues/56)